Fix red highlighting of angle brackets when not part of a bracket. (#214)
# Objective

This extension has the same issue as described
[here](https://github.com/PolyMeilex/vscode-wgsl/issues/20). This PR is
a port of the fix that was eventually applied to `vscode-wgsl`.

## Solution

Remove the angle brackets from the language configuration file. This may
affect highlighting of angle brackets when the cursor is on one side of
a pair of angle brackets, but the red highlighting of other uses of
angle brackets (return syntax, comparison operations) will no longer be
present.