add edition crate (#243)
# Objective

Add a crate to provide an enum for "editions" of the language being
analyzed.
This can be expanded upon and used to support different versions of the
WGSL specification (maybe once stability is a major concern) and also to
support extension to the language such as WESL and extensions thereupon.

This will also make it simpler to port and update other crates from r-a
because there is something to fill that gap, even if it isn't very
useful at the moment.

The first crate that will use this is an update to `crates/parser`.

`xtask` may also use it in the future in order to generate documentation
about lints (after wgsl lints become a feature)

Contributes to #182 

## Solution

Port edition crate from r-a

## Testing

N/A