Fix go to definition in imports (#240)
# Objective

Fixes #147 

## Solution

No idea, I asked claude to do it and it seemed to work. Maintainers
should verify that changes make sense and tweak if necessary

## Testing

- Tested on 0.9.8 on windows and macos in cursor and everything worked
(go to def in imports, type hints, etc).
- Tried testing on latest main (0.11.0) and the go to def in imports
works, but nothing else does (type hints, etc). But I think this is
unrelated to my changes because I can't get the latest main to work
properly even on a fresh clone. Possibly related to #233.
- Definitely needs more testing.

---------

Co-authored-by: Benjamin Brienen <benjamin.brienen@outlook.com>