Use local dependencies instead of workspace dependencies (#519)
# Objective

Dependabot cannot keep workspace dependencies up to date, and we want
Dependabot to keep our dependencies up to date.

https://github.com/dependabot/dependabot-core/issues/7896

## Solution

This PR works around that by not using workspace dependencies in most
cases. The versions will naturally be in sync because we have `cargo
deny` for denying multiple versions of the same dependency in the tree
and Dependabot will update all files at once.

Note that this can be undone using
https://mainmatter.com/blog/2024/03/18/cargo-autoinherit/ in the future
when the Dependabot bug is fixed.