Cargo shear (#742)
# Objective

https://github.com/Boshen/cargo-shear seems to work better than
cargo-udeps. It should also be faster in CI.

## Solution

Switching the CI to it, and removing most of the unused dependencies.
I left some in the root Cargo.toml to test the CI, and also to get
feedback on what to do with them.

## Testing

If it compiles....

## Showcase

Look at how quick our CI for "check-unused-dependencies" now is:
1. 38 seconds in the first run
https://github.com/wgsl-analyzer/wgsl-analyzer/actions/runs/19632317290/job/56214770211?pr=742
2. 30 seconds for the second run
https://github.com/wgsl-analyzer/wgsl-analyzer/actions/runs/19632357001/job/56214934133?pr=742